+++ CHANGELOG.md
@@ -5,8 +5,6 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-
 ## [1.2.0] - TBD
 
 ### Added
@@ -20,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Test suites for the newly added string constraints:
     - `BeEmptyTest` for testing the `BeEmpty` constraint.
     - `ContainTest` for testing the `Contain` constraint.
-
+- `BeMonotonicallyIncreasing` constraint for validating collections of elements to ensure they are monotonically increasing. This constraint is particularly useful in contexts such as sorted lists or sequences where the order of elements is crucial.
 
 ### Removed
 - Pair constraints (`BeFinite.kt`, `BeStrictlyOrdered.kt`, `PairConstraint.kt`) and their corresponding tests. These constraints were found to be obsolete or redundant.